+++ without_bonus/ft_printf.h
@@ -6,7 +6,7 @@
 /*   By: alycgaut <alycgaut@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/11/19 17:17:33 by alycgaut          #+#    #+#             */
-/*   Updated: 2022/12/08 19:29:34 by alycgaut         ###   ########.fr       */
+/*   Updated: 2022/12/09 16:59:37 by alycgaut         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -28,6 +28,6 @@ int		put_char(unsigned int c);
 int		put_str(char *s);
 int		put_ptr(void *ptr_add);
 int		put_uns(unsigned int n);
-int		put_hexa(long long int hex, char c);
+int		put_hexa(unsigned int hex, char c);
 
 #endif
